pyproject.toml
[tool.poetry]
name = "cityweather"
version = "0.1.0"
description = "API for providing weather data by city name"
authors = ["DanTheMain <Daniil.Morozov80@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
httpx = "^0.25.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
faker = "^19.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
